feat: add bookCard in detailProduct Page
